requirement.txt
baidu-aip==2.2.0.0
bs4==0.0.1
